use generic lexer instead of regex lexer
