+++ tests/ParserTests/lexer/GenericLexerTests.cs
@@ -13,6 +13,7 @@ using sly.lexer;
 using sly.parser;
 using sly.parser.generator;
 using Xunit;
+using ExpressionToken = simpleExpressionParser.ExpressionToken;
 
 namespace ParserTests.lexer
 {
